fix: expose codex orchestrator backend controls
